requirements.txt
python-dotenv==1.0.1
discord.py==2.3.2
PyNaCl==1.5.0
python-dateutil==2.8.2